requirements.txt
pandas
numpy
faker
pyyaml
plotly
kaleido
openpyxl